docs(calendar): update calendar component docs
