+++ packages/components/calendar/src/use-month-picker.ts
@@ -1,43 +1,14 @@
-import type {
-  HTMLUIProps,
-  ThemeProps,
-  UIPropGetter,
-  RequiredUIPropGetter,
-} from "@yamada-ui/core"
-import { layoutStyleProperties, useTheme } from "@yamada-ui/core"
-import {
-  formControlProperties,
-  useFormControlProps,
-} from "@yamada-ui/form-control"
-import type { FormControlOptions } from "@yamada-ui/form-control"
-import { popoverProperties, type PopoverProps } from "@yamada-ui/popover"
+import type { UIPropGetter } from "@yamada-ui/core"
 import { useControllableState } from "@yamada-ui/use-controllable-state"
-import { useOutsideClick } from "@yamada-ui/use-outside-click"
-import type { Dict } from "@yamada-ui/utils"
-import {
-  isActiveElement,
-  useUpdateEffect,
-  dataAttr,
-  getEventRelatedTarget,
-  handlerAll,
-  isContains,
-  mergeRefs,
-  pickObject,
-  splitObject,
-  omitObject,
-} from "@yamada-ui/utils"
+import { isActiveElement, useUpdateEffect, handlerAll } from "@yamada-ui/utils"
 import dayjs from "dayjs"
-import type {
-  ChangeEvent,
-  CSSProperties,
-  FocusEvent,
-  KeyboardEvent,
-  MouseEvent,
-} from "react"
-import { useCallback, useRef, useState } from "react"
-import type { CalendarBaseProps } from "./calendar"
-import { isAfterDate, isBeforeDate } from "./calendar-utils"
+import type { ChangeEvent, CSSProperties } from "react"
+import { useCallback, useState } from "react"
 import type { UseCalendarProps } from "./use-calendar"
+import {
+  useCalendarPicker,
+  type UseCalendarPickerProps,
+} from "./use-calendar-picker"
 
 type CalendarProps = Pick<
   UseCalendarProps<Date | undefined>,
@@ -57,270 +28,91 @@ type CalendarProps = Pick<
   | "withLabel"
 >
 
-type CalendarThemeProps = ThemeProps<"Calendar">
+type UseMonthPickerOptions = {
+  /**
+   * The type of the month picker.
+   */
+  type?: "month" | "year"
+  /**
+   * The initial type of the month picker.
+   *
+   * @default 'month'
+   */
+  defaultType?: "month" | "year"
+  /**
+   * The callback invoked when type state changes.
+   */
+  onChangeType?: (type: "month" | "year") => void
+  /**
+   * The format used for conversion.
+   * Check the docs to see the format of possible modifiers you can pass.
+   *
+   * @see Docs https://day.js.org/docs/en/display/format#list-of-localized-formats
+   * @default 'YYYY/MM'
+   */
+  inputFormat?: string
+  /**
+   * If `true`, the list element will be closed when value is selected.
+   *
+   * @default true
+   */
+  closeOnSelect?: boolean
+}
 
-type UseMonthPickerBaseProps = Omit<
-  PopoverProps,
-  | "initialFocusRef"
-  | "closeOnButton"
-  | "isOpen"
-  | "trigger"
-  | "autoFocus"
-  | "restoreFocus"
-  | "openDelay"
-  | "closeDelay"
-> &
-  FormControlOptions &
-  CalendarProps & {
-    /**
-     * The type of the month picker.
-     */
-    type?: "month" | "year"
-    /**
-     * The initial type of the month picker.
-     *
-     * @default 'month'
-     */
-    defaultType?: "month" | "year"
-    /**
-     * The callback invoked when type state changes.
-     */
-    onChangeType?: (type: "month" | "year") => void
-    /**
-     * The pattern used to check the <input> element's.
-     *
-     * @default '/[^0-9\-\/]/g'
-     */
-    pattern?: RegExp
-    /**
-     * Function that converts the input value to Date type.
-     */
-    parseDate?: (value: string) => Date | undefined
-    /**
-     * The format used for conversion.
-     * Check the docs to see the format of possible modifiers you can pass.
-     *
-     * @see Docs https://day.js.org/docs/en/display/format#list-of-localized-formats
-     * @default 'YYYY/MM'
-     */
-    inputFormat?: string
-    /**
-     * If `true`, display the month picker clear icon.
-     *
-     * @default true
-     */
-    isClearable?: boolean
-    /**
-     * If `true`, the list element will be closed when value is selected.
-     *
-     * @default true
-     */
-    closeOnSelect?: boolean
-    /**
-     * If `true`, allows input.
-     *
-     * @default true
-     */
-    allowInput?: boolean
-    /**
-     * Props for calendar component.
-     */
-    calendarProps?: CalendarBaseProps
-  }
-
-export type UseMonthPickerProps = Omit<
-  HTMLUIProps<"input">,
-  | keyof UseMonthPickerBaseProps
-  | "disabled"
-  | "required"
-  | "readOnly"
-  | "size"
-  | "type"
-> &
-  UseMonthPickerBaseProps & {
-    calendarVariant?: CalendarThemeProps["variant"]
-    calendarSize?: CalendarThemeProps["size"]
-    calendarColorScheme?: CalendarThemeProps["colorScheme"]
-  }
+export type UseMonthPickerProps = UseCalendarPickerProps<CalendarProps> &
+  UseMonthPickerOptions
 
 export const useMonthPicker = ({
-  parseDate,
-  defaultIsOpen,
-  closeOnBlur = true,
-  placement = "bottom-start",
-  duration = 0.2,
+  value: valueProp,
   defaultValue,
+  onChange: onChangeProp,
+  type: typeProp,
   defaultType = "month",
-  month,
-  defaultMonth,
-  onChangeMonth,
-  minDate,
-  maxDate,
-  locale,
-  inputFormat = "YYYY/MM",
-  yearFormat,
-  monthFormat,
-  calendarVariant,
-  calendarSize,
-  calendarColorScheme,
-  withHeader,
-  withControls,
-  withLabel,
-  calendarProps,
+  onChangeType: onChangeTypeProp,
   placeholder,
   closeOnSelect = true,
-  allowInput = true,
-  pattern = /[^0-9\-\/]/g,
   ...rest
 }: UseMonthPickerProps) => {
-  const { theme } = useTheme()
-
-  rest = useFormControlProps(rest)
-
-  locale ??= theme.__config.date?.locale ?? "en"
-
-  const { id } = rest
-
-  const formControlProps = pickObject(rest, formControlProperties)
-  const [containerProps, inputProps] = splitObject<Dict, string>(
-    omitObject(rest, [
-      ...popoverProperties,
-      "id",
-      "value",
-      "onChange",
-      "type",
-      "onChangeType",
-    ]),
-    layoutStyleProperties,
-  )
-
-  const stringToDate = useCallback(
-    (value: string): Date | undefined => {
-      let date = parseDate
-        ? parseDate(value)
-        : dayjs(value, inputFormat, locale).toDate()
-
-      if (date == null) return undefined
-
-      if (maxDate && isAfterDate(date, maxDate)) date = maxDate
-      if (minDate && isBeforeDate(date, minDate)) date = minDate
-
-      return date
-    },
-    [inputFormat, locale, maxDate, minDate, parseDate],
-  )
-
-  const dateToString = useCallback(
-    (value: Date | undefined): string | undefined => {
-      if (value == null) return undefined
-
-      if (maxDate && isAfterDate(value, maxDate)) value = maxDate
-      if (minDate && isBeforeDate(value, minDate)) value = minDate
-
-      return dayjs(value)
-        .locale(locale ?? "en")
-        .format(inputFormat)
-    },
-    [inputFormat, locale, maxDate, minDate],
-  )
-
-  const [isOpen, setIsOpen] = useState<boolean>(defaultIsOpen ?? false)
   const [value, setValue] = useControllableState<Date | undefined>({
-    value: rest.value,
+    value: valueProp,
     defaultValue,
-    onChange: rest.onChange,
+    onChange: onChangeProp,
   })
   const [type, setType] = useControllableState({
-    value: rest.type,
+    value: typeProp,
     defaultValue: defaultType,
-    onChange: rest.onChangeType,
+    onChange: onChangeTypeProp,
   })
-  const [inputValue, setInputValue] = useState<string | undefined>(
-    dateToString(value),
-  )
 
-  const containerRef = useRef<HTMLDivElement>(null)
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  const onOpen = useCallback(() => {
-    if (formControlProps.disabled || formControlProps.readOnly) return
-
-    setIsOpen(true)
-
-    if (allowInput && inputRef.current) inputRef.current.focus()
-
-    rest.onOpen?.()
-  }, [allowInput, formControlProps, rest])
-
-  const onClose = useCallback(() => {
-    setIsOpen(false)
-
-    inputRef.current?.blur()
-
-    rest.onClose?.()
-  }, [rest])
-
-  const onClear = useCallback(
-    (ev: MouseEvent<HTMLDivElement>) => {
-      ev.stopPropagation()
-
-      setValue(undefined)
-      setInputValue(undefined)
-    },
-    [setValue],
-  )
-
-  const onClick = useCallback(() => {
-    if (isOpen) {
-      if (allowInput && inputRef.current) inputRef.current.focus()
-    } else {
-      onOpen()
-    }
-  }, [allowInput, isOpen, onOpen])
-
-  const onFocus = useCallback(() => {
-    if (!isOpen) onOpen()
-  }, [isOpen, onOpen])
-
-  const onBlur = useCallback(
-    (ev: FocusEvent<HTMLDivElement>) => {
-      const relatedTarget = getEventRelatedTarget(ev)
-
-      if (isContains(containerRef.current, relatedTarget)) return
-
-      if (!closeOnBlur) return
-
-      const inputValue = dateToString(value)
-
-      setInputValue(inputValue)
-
-      if (isOpen) onClose()
-    },
-    [closeOnBlur, dateToString, isOpen, onClose, value],
-  )
-
-  const onKeyDown = useCallback(
-    (ev: KeyboardEvent<HTMLDivElement>) => {
-      if (formControlProps.disabled || formControlProps.readOnly) return
-
-      const actions: Record<string, Function | undefined> = {
-        Enter: onClose,
-      }
-
-      const action = actions[ev.key]
-
-      if (!action) return
-
-      ev.preventDefault()
-      ev.stopPropagation()
-
-      action(ev)
-    },
-    [formControlProps, onClose],
-  )
-
-  const onChangeType = useCallback(
-    (type: "month" | "year" | "date", year?: number, month?: number) => {
+  const {
+    inputRef,
+    id,
+    allowInput,
+    pattern,
+    inputProps,
+    formControlProps,
+    onClose,
+    dateToString,
+    stringToDate,
+    getContainerProps,
+    getPopoverProps,
+    getFieldProps,
+    getCalendarProps,
+    getIconProps,
+  } = useCalendarPicker({
+    inputFormat: "YYYY/MM",
+    ...rest,
+    selectMonthWith: "value",
+    value,
+    defaultValue,
+    onChange: setValue,
+    type,
+    defaultType,
+    onChangeType: (
+      type: "month" | "year" | "date",
+      year?: number,
+      month?: number,
+    ) => {
       if (type !== "date") {
         setType(type)
       } else {
@@ -337,10 +129,24 @@ export const useMonthPicker = ({
         if (closeOnSelect && value) onClose()
       }
     },
-    [closeOnSelect, dateToString, onClose, setType, setValue],
+    onClear: () => {
+      setValue(undefined)
+      setInputValue(undefined)
+    },
+    onClose: () => {
+      const inputValue = dateToString(value)
+
+      setInputValue(inputValue)
+
+      rest.onClose?.()
+    },
+  })
+
+  const [inputValue, setInputValue] = useState<string | undefined>(
+    dateToString(value),
   )
 
-  const onInputChange = useCallback(
+  const onChange = useCallback(
     (ev: ChangeEvent<HTMLInputElement>) => {
       let inputValue = ev.target.value
 
@@ -367,61 +173,6 @@ export const useMonthPicker = ({
     setInputValue(inputValue)
   }, [value])
 
-  useOutsideClick({
-    ref: containerRef,
-    handler: onClose,
-    enabled: closeOnBlur,
-  })
-
-  const getContainerProps: UIPropGetter = useCallback(
-    (props = {}, ref = null) => ({
-      ref: mergeRefs(containerRef, ref),
-      ...containerProps,
-      ...props,
-      ...formControlProps,
-      onClick: handlerAll(props.onClick, rest.onClick, onClick),
-      onBlur: handlerAll(props.onBlur, rest.onBlur, onBlur),
-    }),
-    [containerProps, formControlProps, onBlur, onClick, rest],
-  )
-
-  const getPopoverProps = useCallback(
-    (props?: PopoverProps): PopoverProps => ({
-      ...rest,
-      ...props,
-      isOpen,
-      onOpen,
-      onClose,
-      placement,
-      duration,
-      trigger: "never",
-      closeOnButton: false,
-    }),
-    [duration, isOpen, onClose, onOpen, placement, rest],
-  )
-
-  const getFieldProps: UIPropGetter = useCallback(
-    (props = {}, ref = null) => {
-      const style: CSSProperties = {
-        ...props.style,
-        ...(!allowInput ? { cursor: "pointer" } : {}),
-      }
-
-      return {
-        ref: mergeRefs(ref, inputRef),
-        tabIndex: !allowInput ? 0 : -1,
-        ...props,
-        ...formControlProps,
-        style,
-        "data-active": dataAttr(isOpen),
-        "aria-expanded": dataAttr(isOpen),
-        onFocus: handlerAll(props.onFocus, rest.onFocus, onFocus),
-        onKeyDown: handlerAll(props.onKeyDown, rest.onKeyDown, onKeyDown),
-      }
-    },
-    [allowInput, formControlProps, isOpen, rest, onFocus, onKeyDown],
-  )
-
   const getInputProps: UIPropGetter = useCallback(
     (props = {}) => {
       const style: CSSProperties = {
@@ -441,7 +192,7 @@ export const useMonthPicker = ({
         value: inputValue ?? "",
         cursor: formControlProps.readOnly ? "default" : "text",
         pointerEvents: formControlProps.disabled ? "none" : "auto",
-        onChange: handlerAll(props.onChange, onInputChange),
+        onChange: handlerAll(props.onChange, onChange),
       }
     },
     [
@@ -451,74 +202,9 @@ export const useMonthPicker = ({
       formControlProps,
       id,
       inputValue,
-      onInputChange,
+      onChange,
     ],
   )
-
-  const getCalendarProps = useCallback(
-    (
-      props?: CalendarProps,
-    ): CalendarProps & {
-      type: "month" | "year"
-      onChangeType: UseCalendarProps["onChangeType"]
-      selectMonthWith: UseCalendarProps["selectMonthWith"]
-    } => ({
-      ...props,
-      type,
-      onChangeType,
-      month,
-      defaultMonth,
-      onChangeMonth,
-      yearFormat,
-      monthFormat,
-      withHeader,
-      withControls,
-      withLabel,
-      variant: calendarVariant,
-      size: calendarSize,
-      colorScheme: calendarColorScheme,
-      ...calendarProps,
-      value,
-      defaultValue,
-      onChange: setValue,
-      minDate,
-      maxDate,
-      locale,
-      selectMonthWith: "value",
-    }),
-    [
-      calendarColorScheme,
-      calendarProps,
-      calendarSize,
-      calendarVariant,
-      defaultMonth,
-      defaultValue,
-      locale,
-      maxDate,
-      minDate,
-      month,
-      monthFormat,
-      onChangeMonth,
-      onChangeType,
-      setValue,
-      type,
-      value,
-      withControls,
-      withHeader,
-      withLabel,
-      yearFormat,
-    ],
-  )
-
-  const getIconProps: RequiredUIPropGetter<"div", { clear: boolean }> =
-    useCallback(
-      ({ clear, ...props }) => ({
-        ...props,
-        ...formControlProps,
-        onClick: handlerAll(props.onClick, clear ? onClear : undefined),
-      }),
-      [formControlProps, onClear],
-    )
 
   return {
     value,
